+++ twitter4j-core/src/main/java/twitter4j/UserStreamAdapter.java
@@ -31,7 +31,7 @@ package twitter4j;
  * @since Twitter4J 2.1.3
  */
 public class UserStreamAdapter extends StatusAdapter implements UserStreamListener{
-    public void onDeletionNotice(int directMessageId, int userId) {
+    public void onDeletionNotice(long directMessageId, int userId) {
     }
 
     public void onFriendList(int[] friendIds) {
